mod: use rounded value for accuracy and HS stats for now, refs #478

+++ src/cgame/cg_local.h
@@ -583,8 +583,8 @@ typedef struct clientInfo_s
 	qboolean selected;
 
 	// Intermission stats
-	float totalWeapAcc;
-	float totalWeapHSpct;
+	int totalWeapAcc;
+	int totalWeapHSpct;
 	int kills;
 	int deaths;
 	int selfkills;
